\name{shinyIRT}
\alias{shinyIRT}
\title{Item Response Theory Analysis with a Shiny Application}
\usage{
shinyIRT()
}
\value{No return value, launches 'shiny' application}
\description{
Performing Item Response Thory analysis such as paramater estimation, ability estimation, item and model fit analyse, local independence assumption, dimensionality assumption, charachteristic and information curves under various models with a user friendly 'shiny' interface.
}
\examples{
if(interactive()){
## Run this code for launching the 'shiny' application
shinyIRT()
}
}
